// types.ts
export interface BaseItem {
  id: string;
  name: string;
  description?: string;
}
// ...
export interface CascadingOption extends BaseItem {
  parentId: string | null;
  jsonBlock?: string;
  description?: string;
}